Remove `go func` parallelization in ABCI EndBlocker (#440)
## What is the purpose of the change

This change simplifies a hard-to-read and hard-to-reason about piece of
code that parallelized the picking of churn ready topics from the active
set. After this PR the code is easier to read and understand, and has
less likelihood of causing non-determinism issues. The tradeoff is that
the code may run slightly slower in serial.

## Testing and Verifying

This change is not trivial, however it is a small diff / refactor PR and
our existing test cases should be able to cover the change.

## Documentation and Release Note

This PR does not have any implications for documentation.